requirements.txt
Flask==3.0.0
pdfplumber==0.10.3
groq==0.9.0
python-dotenv==1.0.0
httpx==0.24.1
PyPDF2==3.0.1
pdfminer.six